refactor: split CommandManager into focused sub-modules
Split the monolithic CommandManager.ts (1516 lines) into 6 cohesive
modules while keeping a thin orchestrator for event listeners and
sub-module instantiation. No behavioral changes.
